o.a.p.c.schema -> o.a.p.c.protocol.schema
Classes in pulsar-common module were moved to avoid jpms split-package
conflict between the pulsar-client-api and pulsar-common modules.
