+++ pulsar-common/src/main/java/org/apache/pulsar/common/protocol/schema/DeleteSchemaResponse.java
@@ -16,23 +16,17 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.common.schema;
+package org.apache.pulsar.common.protocol.schema;
 
-import java.util.HashMap;
-import java.util.Map;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.ToString;
+import lombok.NoArgsConstructor;
 
-@Builder
 @Data
-@ToString
-public class SchemaData {
-    private final SchemaType type;
-    private final boolean isDeleted;
-    private final long timestamp;
-    private final String user;
-    private final byte[] data;
-    @Builder.Default
-    private Map<String, String> props = new HashMap<>();
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class DeleteSchemaResponse {
+    private long version;
 }
