+++ pulsar-common/src/main/java/org/apache/pulsar/common/protocol/schema/GetSchemaResponse.java
@@ -16,17 +16,23 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.common.schema;
+package org.apache.pulsar.common.protocol.schema;
 
+import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.apache.pulsar.common.schema.SchemaType;
 
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class PostSchemaResponse {
-    private SchemaVersion version;
+public class GetSchemaResponse {
+    private long version;
+    private SchemaType type;
+    private long timestamp;
+    private String data;
+    private Map<String, String> properties;
 }
