+++ pulsar-common/src/main/java/org/apache/pulsar/common/protocol/schema/SchemaInfoUtil.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.common.schema;
+package org.apache.pulsar.common.protocol.schema;
 
 import java.util.Collections;
 import java.util.TreeMap;
@@ -26,6 +26,7 @@ import lombok.experimental.UtilityClass;
 import org.apache.pulsar.common.protocol.Commands;
 import org.apache.pulsar.common.api.proto.PulsarApi;
 import org.apache.pulsar.common.api.proto.PulsarApi.Schema;
+import org.apache.pulsar.common.schema.SchemaInfo;
 
 @UtilityClass
 public class SchemaInfoUtil {
